+++ lib/td_dq/accounts/user.ex
@@ -1,4 +1,6 @@
 defmodule TdDq.Accounts.User do
   @moduledoc false
+
+  @derive Jason.Encoder
   defstruct id: 0, user_name: nil, is_admin: false, jti: nil
 end