+++ lib/td_dq/auth/error_handler.ex
@@ -2,8 +2,10 @@ defmodule TdDq.Auth.ErrorHandler do
   @moduledoc false
   import Plug.Conn
 
+  alias Jason, as: JSON
+
   def auth_error(conn, {type, _reason}, _opts) do
-    body = Poison.encode!(%{message: to_string(type)})
+    body = JSON.encode!(%{message: to_string(type)})
     send_resp(conn, 401, body)
   end
 end