+++ lib/td_dq/search/indexer.ex
@@ -1,7 +1,9 @@
 defmodule TdDq.Search.Indexer do
   @moduledoc """
-    Manages elasticsearch indices
+  Manages elasticsearch indices
   """
+
+  alias Jason, as: JSON
   alias TdDq.ESClientApi
   alias TdDq.Search
   alias TdDq.Search.RuleMappings
@@ -9,7 +11,7 @@ defmodule TdDq.Search.Indexer do
   def reindex(index_name, items) do
     ESClientApi.delete!(index_name)
     mapping = RuleMappings.get_mappings()
-    %{status_code: 200} = ESClientApi.put!(index_name, mapping |> Poison.encode!())
+    %{status_code: 200} = ESClientApi.put!(index_name, mapping |> JSON.encode!())
     Search.put_bulk_search(items)
   end
 end