+++ lib/td_dq_web/endpoint.ex
@@ -19,7 +19,7 @@ defmodule TdDqWeb.Endpoint do
   plug(Plug.Parsers,
     parsers: [:urlencoded, :multipart, :json],
     pass: ["*/*"],
-    json_decoder: Poison
+    json_decoder: Jason
   )
 
   plug(Plug.MethodOverride)