+++ test/support/rule.ex
@@ -1,10 +1,11 @@
 defmodule TdDqWeb.Rule do
   @moduledoc false
 
-  alias Poison, as: JSON
   import TdDqWeb.Router.Helpers
   import TdDqWeb.Authentication, only: :functions
   import TdDqWeb.SupportCommon, only: :functions
+
+  alias Jason, as: JSON
 
   @endpoint TdDqWeb.Endpoint