+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - [TD-1618] Cache improvements (use td-cache instead of td-perms)
+- [TD-1924] Use Jason instead of Poison for JSON encoding/decoding
 
 ## [3.0.0] 2019-06-25